Move controller-related items into package
